Added PID library and Top.h for constant centralization, also changed a lot of variables to floats instead of doubles to save memory

+++ kukie/Encoder/Encoder.h
@@ -16,7 +16,7 @@ public:
     Encoder(int wheelDiameterMM);
     unsigned long getPrevTime();
     unsigned long setTime(unsigned long t);
-    double getSpeed(unsigned long tMicros);
+    float getSpeed(unsigned long tMicros);
 private:
     int _dia;
     unsigned long _prevtime;
